+++ modules/hanna-react/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- `Alert`:
+  - feat: Auto-show close button if `onClosed` is set without `autoClose`
 
 ## 0.10.164